pre switching to completable
