+++ app/src/main/java/com/lightricks/feedexercise/application/AppContainer.kt
@@ -1,0 +1,10 @@
+package com.lightricks.feedexercise.application
+
+import android.content.Context
+import com.lightricks.feedexercise.database.FeedDatabase
+
+class AppContainer(context: Context) {
+
+    val database = FeedDatabase.getDatabase(context)
+
+}
