+++ app/src/main/java/com/lightricks/feedexercise/application/FeedExercise.kt
@@ -1,0 +1,7 @@
+package com.lightricks.feedexercise.application
+
+import android.app.Application
+
+class FeedExercise : Application() {
+    val appContainer = AppContainer(this)
+}
